Stores to project folder, remembers settings.
The app now remembers the settings from last time. Users can name the
project and the cameras. The app will create a folder for the project
and a file for each camera. Files are numbered, so when you press
"Stop" and then "Record" again, your previous recording does not
get overwritten.

I seem to have broken something - The software still works on my dev
machine, but does nothing on my recoring computer. Seems like I have
a task for tomorrow ;)